feat(FRO-X): converted signup page to a reusable component
